Refactor: move ColonyGrid to shared
