+++ src/components/shared/ColonyGrid/ColonyGridItem.css.d.ts
